checkmate/backend/authLogin/requirements.txt
bcrypt==4.2.0
boto3==1.35.14
PyJWT==2.9.0
